Fix float64 upcast from np.gradient in build_pitch_features
np.gradient returns float64 regardless of input dtype, causing the stacked
pitch array to upcast to float64 and break Conv1d (expects float32).

+++ voice_restore/train.py
@@ -219,7 +219,7 @@ def build_pitch_features(f0_np: np.ndarray,
     delta     = np.gradient(f0_norm)
     delta2    = np.gradient(delta)
 
-    return np.stack([f0_norm, conf, delta, delta2], axis=0)   # (4, T)
+    return np.stack([f0_norm, conf, delta, delta2], axis=0).astype(np.float32)   # (4, T)
 
 
 def build_harmonic_features(f0_np: np.ndarray, n_frames: int) -> np.ndarray:
